--- src/types/pagination.ts
@@ -1,6 +1,0 @@
-export interface PageInfo {
-    page: number;
-    size: number;
-    totalPages: number;
-    totalCount: number;
-}
